Move CloseButton to molecules
