+++ src/ui/layout/navbar-header.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { type Snippet } from "svelte";
   import ModalSection from "./modal-section.svelte";
-  import CloseButton from "../components/close-button.svelte";
+  import CloseButton from "../molecules/close-button.svelte";
   import BackButton from "../components/back-button.svelte";
   import IconArrow from "../atoms/icons/icon-arrow.svelte";
   import { LocalizationKeys } from "../localization/supportedLanguages";
